+++ apps/mobile-wallet/src/features/backup/backupScreens/VerifyMnemonicScreen.tsx
@@ -20,7 +20,8 @@ import SecretPhraseWordList, { SelectedWord } from '~/components/SecretPhraseWor
 import { useHeaderContext } from '~/contexts/HeaderContext'
 import { useAppDispatch, useAppSelector } from '~/hooks/redux'
 import { BackupMnemonicNavigationParamList } from '~/navigation/BackupMnemonicNavigation'
-import { dangerouslyExportWalletMnemonic, updateStoredWalletMetadata } from '~/persistent-storage/wallet'
+import { updateStoredWalletMetadata } from '~/persistent-storage/wallet'
+import { dangerouslyExportWalletMnemonic } from '~/persistent-storage/walletMnemonic'
 import { PossibleWordBox, SecretPhraseBox, Word } from '~/screens/new-wallet/ImportWalletSeedScreen'
 import { mnemonicBackedUp } from '~/store/wallet/walletSlice'
 import { DEFAULT_MARGIN } from '~/style/globalStyle'
