+++ apps/mobile-wallet/src/persistent-storage/wallet.ts
@@ -1,20 +1,13 @@
-import {
-  dangerouslyConvertUint8ArrayMnemonicToString,
-  keyring,
-  mnemonicJsonStringifiedObjectToUint8Array
-} from '@alephium/keyring'
+import { keyring } from '@alephium/keyring'
 import {
   AddressHash,
   addressMetadataIncludesHash,
   AddressStoredMetadataWithHash,
-  DeprecatedWalletMetadataMobile,
   GROUPLESS_ADDRESS_KEY_TYPE,
-  resetArray,
   WalletMetadataMobile
 } from '@alephium/shared'
 import { nanoid } from 'nanoid'
 
-import { sendAnalytics } from '~/analytics'
 import { deleteFundPassword } from '~/features/fund-password/fundPasswordStorage'
 import i18n from '~/features/localization/i18n'
 import { deleteAddressKeyPair, generateAndStoreAddressKeypairForIndex } from '~/persistent-storage/addressKeys'
@@ -22,7 +15,6 @@ import { legacyGetWalletMetadata, legacyStoredMnemonicV2Exists } from '~/persist
 import { storage } from '~/persistent-storage/storage'
 import {
   deleteSecurelyWithReportableError,
-  getSecurelyWithReportableError,
   storeSecurelyWithReportableError,
   storeWithReportableError
 } from '~/persistent-storage/utils'
@@ -33,11 +25,11 @@ import {
   removeWalletFromList,
   walletListExists
 } from '~/persistent-storage/walletList'
+import { getStoredWalletMetadata, getWalletMetadata, storeWalletMetadata } from '~/persistent-storage/walletMetadata'
+import { walletMnemonicKey } from '~/persistent-storage/walletMnemonic'
 import { GeneratedWallet, WalletStoredState } from '~/types/wallet'
 import { getRandomLabelColor } from '~/utils/colors'
 
-export const walletMnemonicKey = (walletId: string) => `wallet-mnemonic-${walletId}`
-const walletMetadataKey = (walletId: string) => `wallet-metadata-${walletId}`
 const isNewWalletKey = (walletId: string) => `is-new-wallet-${walletId}`
 
 export const generateAndStoreWallet = async (
@@ -71,47 +63,6 @@ export const generateAndStoreWallet = async (
   }
 }
 
-export const getWalletMetadata = (
-  walletId: string,
-  throwError = true
-): WalletMetadataMobile | DeprecatedWalletMetadataMobile | null => {
-  let rawWalletMetadata
-  let walletMetadata = null
-
-  try {
-    rawWalletMetadata = storage.getString(walletMetadataKey(walletId))
-  } catch (error) {
-    if (throwError) throw error
-  }
-
-  if (rawWalletMetadata) {
-    try {
-      walletMetadata = JSON.parse(rawWalletMetadata)
-    } catch (error) {
-      sendAnalytics({ type: 'error', error, message: 'Could not parse wallet metadata' })
-      if (throwError) throw error
-    }
-  }
-
-  return walletMetadata
-}
-
-export const getStoredWalletMetadata = (
-  walletId: string,
-  error?: string
-): WalletMetadataMobile | DeprecatedWalletMetadataMobile => {
-  const walletMetadata = getWalletMetadata(walletId)
-
-  if (!walletMetadata)
-    throw new Error(error || `${i18n.t('Could not get stored wallet')}: ${i18n.t('Wallet metadata not found')}`)
-
-  return walletMetadata
-}
-
-export const isStoredWalletMetadataMigrated = (
-  metadata: WalletMetadataMobile | DeprecatedWalletMetadataMobile
-): metadata is WalletMetadataMobile => (metadata as WalletMetadataMobile).addresses.every(addressMetadataIncludesHash)
-
 export const updateStoredWalletMetadata = (walletId: string, partialMetadata: Partial<WalletMetadataMobile>) => {
   const walletMetadata = getStoredWalletMetadata(
     walletId,
@@ -133,7 +84,7 @@ export const deleteWallet = async (walletId: string) => {
 
   await deleteSecurelyWithReportableError(walletMnemonicKey(walletId), true, '')
   await deleteFundPassword(walletId)
-  storage.delete(walletMetadataKey(walletId))
+  storage.delete(`wallet-metadata-${walletId}`)
   storage.delete(isNewWalletKey(walletId))
   removeWalletFromList(walletId)
 }
@@ -180,37 +131,6 @@ export const getIsNewWallet = (walletId: string): boolean | undefined => storage
 
 export const storeIsNewWallet = (walletId: string, isNew: boolean) =>
   storeWithReportableError(isNewWalletKey(walletId), isNew)
-
-export const storedMnemonicExists = async (walletId: string): Promise<boolean> =>
-  !!(await getSecurelyWithReportableError(walletMnemonicKey(walletId), true, ''))
-
-export const dangerouslyExportWalletMnemonic = async (walletId: string): Promise<string> => {
-  const decryptedMnemonic = await getSecurelyWithReportableError(walletMnemonicKey(walletId), true, '')
-
-  if (!decryptedMnemonic)
-    throw new Error(`${i18n.t('Could not export mnemonic')}: ${i18n.t('Could not find stored wallet')}`)
-
-  const parsedDecryptedMnemonic = mnemonicJsonStringifiedObjectToUint8Array(JSON.parse(decryptedMnemonic))
-
-  return dangerouslyConvertUint8ArrayMnemonicToString(parsedDecryptedMnemonic)
-}
-
-export const initializeKeyringWithStoredWallet = async (walletId: string) => {
-  let decryptedMnemonic = await getSecurelyWithReportableError(walletMnemonicKey(walletId), true, '')
-  if (!decryptedMnemonic)
-    throw new Error(`${i18n.t('Could not initialize keyring')}: ${i18n.t('Could not find stored wallet')}`)
-
-  const parsedDecryptedMnemonic = mnemonicJsonStringifiedObjectToUint8Array(JSON.parse(decryptedMnemonic))
-  keyring.initFromDecryptedMnemonic(parsedDecryptedMnemonic, '')
-
-  decryptedMnemonic = ''
-  resetArray(parsedDecryptedMnemonic)
-}
-
-export const storeWalletMetadata = (
-  walletId: string,
-  metadata: WalletMetadataMobile | DeprecatedWalletMetadataMobile
-) => storeWithReportableError(walletMetadataKey(walletId), JSON.stringify(metadata))
 
 export const storeWalletMnemonic = async (walletId: string, mnemonic: Uint8Array) =>
   storeSecurelyWithReportableError(walletMnemonicKey(walletId), JSON.stringify(mnemonic), true, '')
