+++ apps/mobile-wallet/src/screens/LoginWithPinScreen.tsx
@@ -12,7 +12,7 @@ import { useAppDispatch, useAppSelector } from '~/hooks/redux'
 import { useBiometrics } from '~/hooks/useBiometrics'
 import RootStackParamList from '~/navigation/rootStackRoutes'
 import { legacyGetStoredWalletMetadata, migrateDeprecatedMnemonic } from '~/persistent-storage/legacyWallet'
-import { isStoredWalletMetadataMigrated } from '~/persistent-storage/wallet'
+import { isStoredWalletMetadataMigrated } from '~/persistent-storage/walletMetadata'
 import { mnemonicMigrated } from '~/store/wallet/walletActions'
 import { showExceptionToast } from '~/utils/layout'
 import { resetNavigation } from '~/utils/navigation'
